tools/hotel-analysis/apps/reporter/pyproject.toml
[project]
name = "reporter"
version = "0.1.0"
description = "Reads assessment JSON from stdin, produces DOCX using template; charts via matplotlib."
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools]
# Flat layout: multiple top-level .py modules by design; list them explicitly.
py-modules = [
    "add_ada_vuln_styles_to_template",
    "constants",
    "dev_smoke",
    "docx_ops",
    "dump_template_styles",
    "inject_anchors",
    "inject_anchors_into_body",
    "io_payload",
    "main",
    "perf_reporter",
    "qc_export",
    "qc_pipeline",
    "render_part1",
    "render_part2",
    "render_vulns",
    "report_humanize",
    "sanitize",
    "validate_template_docx",
    "verify_output",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E402", "I001", "F401"]
